[Place Page Revamp] Refactors place page into using the new json format for the home page. (#4838)
Uses blocks to display the charts in the place page - adds the per
capita checkbox when needed.
Uses a pattern of matching Line charts with highlight charts, Maps with
rankings charts, and uses bar charts for comparisons.

See screencast showing current behavior:
https://screencast.googleplex.com/cast/NjA5MzQyODE3MjUyMTQ3Mnw5ODZjN2RiNS1mZg
Still some bugs to resolve, and more charts need to be configured but
this sets up the foundations.