[Place Page Revamp] Optimization on existence check and force include the place. (#4850)
Optimizes the existence check by always running the existence check on
the full chart config, then filtering out the blocks that are for a
different category. Previously we ran 2 existence checks, on the full
chart config (to know which categories have data) and then on the
category only chart config.

We also remove the bar chart blocks if there is no data for the current
place.

And we now pass in a list of DCIDs as peers within the parent which
returns peer places within a parent to highlight.
That list is used in the comparisonPlaces param in the TileConfig. Note
that we randomize the order of the list on the server side.


https://screencast.googleplex.com/cast/NDk5MDU2ODg4NjE3MzY5NnxkMDVkOWM0YS02Yw