[Place Page Revamp] Filter out bar charts unless we have data for 2+ peers within parents (#4866)
Removes charts like this:
[autopush](https://screenshot.googleplex.com/ocLuP7AYsis8bNW)
[local](https://screenshot.googleplex.com/53yFY92Fr95om3r).

Only keeping charts if we have 2 or more places to compare it with.
For that, I fetch the peers within parent list and trim to the first 15
(that's how many we use for now), and check that at least 2 have data
for each sv. We rely on the cache of fetch_peers_within_parent to
guarantee that the same list will be returned from the related_places
API call and in the filter_charts call within the api call to fetch
charts.