[Rollout Infra] Adds the fallback flags, creates the script to upload to GCS and modifies the loading logic. (#4871)
This PR modifies the logic for reading the feature flags such that it
falls back to local files if for some reason the GCS bucket file is not
found, if there's a permission error, etc.

We also modify the logic on the 3 existing flags (autocomplete, Place
page experiment, Place page GA) to represent what is needed.

We create the local files to be used as fallback, and from which the GCS
files will get updated. Note that we enabled Autocomplete on all
environment (for the next release). But kept Place page experiment and
GA disabled for now.

Creates the script to update the GCS files from the local files.
For production, we also diff the file in the staging GCS bucket in order
to ensure it has been verified.

Lastly the script is able to trigger a Kubernetes restart.

This has all been tested end to end.