[Place Page Revamp] Refactor place page code & use asyncio calls to parallelize expensive operations (#4912)
With some of the recent place page work, the latency on the api calls
has reached levels above what we should. This PR adds parallelization to
our expensive operations with the related_places, and chats endpoints
for the revamped place page.

This also refactors the code significantly to reuse code across
different parts of the functions, and create more granular helper
functions.
In a follow up PR I will add thorough testing on all these helpers and
the main api endpoints.

See Latency with this PR applied:
https://paste.googleplex.com/4715161129320448
See latency without this PR:
https://paste.googleplex.com/5451439923789824
Essentially, this leads to an approximately 40-50% speed up in the
charts endpoint, and 30-40% for the related places endpoint.