[Feature Cleanup] Remove references in the code to the scroll to top button (#4957)
This PR removes references in the code to gate the scroll to top button
behind a feature.
I also mark the feature flag in all environments as deprecated in order
to highlight that it can be removed. But I do not remove it yet because
we need this PR to be included in a release first.

If we remove the flags right now, we would end up disabling the feature
until this PR makes it to production. So we ll keep the flags in for
just a couple more weeks.