Fix flaky embeddings tests (#4919)
Through testing, it looks like the problem with the flaky embeddings
tests arises when running the tests in parallel & seems to have to do
with reading the downloaded files in parallel. Removed the flag that
runs the tests in parallel and have tested and seen that this doesn't
impact the timing of the tests.