[Dynamic Placeholders] Modifies the Placeholder in the search bar to type sample questions (#4974)
We are adding this feature in order to provide users inspiration for the
type of questions they can ask. We use show the characters typed in
order to create some movement on the page and help attract users'
attention.

In order to get this to work, I had to make some modifications such that
the placeholder gets stored in the QueryStore, so that we can safely
override it in the input, then return to it after we have cycled through
a few questions.

This PR only gets this feature to work on the homepage, new place page
and old place page. The explore page always has a query set (not
placeholder). We do NOT override a query in the search bar to show this.

This also adds google analytics tracking, such that all future queries
will show whether they were executed while this feature was on or off.

We added 2 feature flags for this, one for a GA rollout, and one for an
experiment setup with the experimental rollout set at 20% of pages.

We reused the sample queries already displayed in the home page tiles.

Here is a screencast of how it works:

https://screencast.googleplex.com/cast/NTU4OTMxNDQzOTIxNzE1MnxkZGIzZDAwNC1kYg

Note we added a couple seconds delay before it starts, then we cycle
through 5 questions from the sample questions. Then we stop and revert
to the original placeholder.