[Placeholder Tweaks] Change message, delay, styling and theme typing (#4992)
Changes the delay to begin after 0.5 seconds, as requested.
Changes the questions we cycle through, and adds places.
Disables dynamic placeholders on mobile due to screen size and reverts
styling changes done in order to make it work on mobile.
Triggers Google Analytics events accordingly.
Changes the speed.
Fixes the theme typing, still able to start the node server! 

This is a copy of https://github.com/datacommonsorg/website/pull/4980,
with minor edits and the theme typing fix.