Added keyboard functionality to the tooltips.
This required some special handling (and so, a bit of reworking of some aspects of the tooltip).

The functionality is:
- When you tab to a tooltip trigger, the tooltip opens.
- If the tooltip contains no tabbable items itself, the next tab just moves passed the trigger (closing the tooltip).
- If the tooltip contains tabble items, then the next tab goes into the tooltip, and subsequent tabs hit any additional tabbable items. After the final item, the next tab moves to the next item after the trigger outside the tooltip, with the tooltip closing.
- Reverse tabbing (with shift) will do the same but in reverse.

This functionality involved detecting whether the trigger is a tabbable item in its own right. If not, we wrap it in a div that is tabbable. If it is, we clone it and feed it the handlers we need for the tabbable item. By creating a single trigger element with the tab order at the same level, this allows us to have a very consistent tab order for us to be able to move from tooltip to tooltip (or to other elements) with the keyboard.