A fairly significant update of the tooltip.
The primary purpose of the commit is to add buffering to either lateral side of the tooltip (i.e., if the tooltip is above, then there is buffer to the left and the right of the tooltip).

This buffer makes it easier to transition with the mouse from the trigger to the tooltip when the trigger is small (such as with an icon) and so makes the tooltips more friendly.

In making this update, we have also now removed the tooltip container (which was previously providing the safe "bridging" between the trigger and the tooltip) and instead replaced it with a buffering bridge.

We also restored the fade functionality.