The older, simpler metadata modal has been moved to a new component, and we have now added in a new metadata modal for the upcoming more complex metadata modal.
The older metadata modal is still used as a fallback for when we do not have the facet information required to display the more complex version of the modal.

Each chart will need to be updated to supply this information (rather than just the simple list of source strings they are currently sending).

Some minor unrelated updates to typing and JSDoc comments in some of the files are also included.