update nodejs endpoint to be able to specify detector (#5145)
Update the nodejs endpoint to allow users to specify the detector type
for NL detection.

This came up because "What is the most populous state in the US" works
on Data Commons website which uses detector=hybrid, but does not work
through the nodejs endpoint which uses detector=heuristic. We want to
keep the default behavior, but allow a user to specify if they want to
use the hybrid detection.

With this change:
- default:
[/nodejs/query?q=what%20is%20the%20most%20populous%20state%20in%20the%20us](https://screenshot.googleplex.com/4B6nP8cqA5JznGD)
- with hybrid detection:
[/nodejs/query?q=what%20is%20the%20most%20populous%20state%20in%20the%20us&detector=hybrid](https://screenshot.googleplex.com/3ofpz3qK6yHL2mo)