Added in a two new utility funtions.
truncateString is a generic function to truncate strings and replace the truncation with an elliptical string (by default an ellipsis but this can be specified). The string can truncate either at the beginning, middle or end of the string.

A second utility function trims the protocol off URLs. This is a fairly common task, and is done by `urlToDisplayText` (which we used to use). However, we don't want to do everything that urlToDisplayText does, and stripping off the protocol may be a fairly common task on its own.

The truncation of the source URL is then implemented in the metadata modal content.