Metadata Dialog (#5147)
## Description

This PR implements the new metadata modal for charts. The new metadata
modal content is substantially different than before, where only a list
of included stat vars was provided. The metadata now retrieved and
displayed contains information from the details of each stat var, the
details of each source as well as specific details pulled from the
knowledge graph via their DCIDs. Because of the complexity of the
information being pulled, this will be outlined in detail below.

## Notes

### Required Information.

Before this PR, only the stat vars and the source URLs were being passed
up from each chart to `<TileSources> `component (where the metadata
modal is triggered). This is because the old metadata modal only needed
the stat var list, and `<TileSources>` then displayed those source URLs
in a list.

However, in order to retrieve the information now displayed in the new
metadata modal, we need: the stat vars, the sources themselves (the
dcids of the sources) and a map of which sources are used by which stat
var.

This information is available at each chart, but was not being collated
and sent up to other components. So the first task of this PR was to
retrieve all of that information from each chart, and send it up to
where the metadata modals can use it. This accounts for changes in files
such as `bar_tile.tsx`, `line_tile.tsx`, and `chart_tile.tsx`, among
others. These changes are all to pass the needed facet information
(which gives us the sources) to them metadata dialog.

A note that we did apply this to most of the charts, but there were some
charts we could see and therefore verify the functionality of within the
system. If a chart does not pass the required metadata information up,
it will default to use the old stat var only metadata.

### Refactoring

Because of the increased complexity of these components, the TileSources
from `tile_utils` has been moved into its own components and metadata
functionality is now in its own location. While working in
tile_utils.tsx, we also updated some of the code to remove
TypeScript/ESLint warnings. We’ve done this in a few files that we
touched.

### The API Calls

The API calls (located in `<TileMetadataModal />` are documented, but to
outline the process here:

1. First we retrieve the full stat var names (via the dcids that we
have). This was being done in the old modal as well.
2. We then make the necessary calls to get the full category names for
each stat vars (i.e., “Demographics”).
3. We then pull the full stat var information for each stat var (that
gives us details for periodicity, dates, measurement methods, etc for
particular stat var / source combinations)
4. We then pull the full source (provenance) data, in order to get
things like the source name and license type.
5. Finally we look up any measurement methods we might have, converting
the DCIDs to the description.

This all happens quickly, but in a future PR we may look at combining
these calls into a single call.

### The Metadata Content

The metadata content is then rendered with the the labels and field name
translated.

#### Periodicity (Publication Cadence).

Periodicity comes as an ISO 8106 duration (i.e. “P1D” for 1 day). To
facilitate the conversion of this to natural language, we built a
conversion function (which converts it into text that makes sense for
the concept of cadence) and we created a set of translation methods
(located in `i18n_periodicity_messages` to facilitate translation of
these to other languages. This leverages ICU message formats for
handling plurals and combinations. These provide a flexible way for
translators to convert strings like this into other languages where
pluralization is very different. This could be something we use in other
places as well.

We also provided a test file for the conversion function (currently only
testing the English results).

#### Metadata fields

To help understand where the metadata information comes from, attached
is a screenshot of an example metadata field, and corresponding
screenshots mapping where the data comes from (with annotations below
where relevant).

<img width="918" alt="metadata-example"
src="https://github.com/user-attachments/assets/e26a98d1-9ea2-41a5-a5f1-1ca6e24a92d4"
/>
<img width="1074" alt="stat-var"
src="https://github.com/user-attachments/assets/9921c0cf-c0e3-4037-9f7a-51a1bdb891db"
/>
<img width="983" alt="source"
src="https://github.com/user-attachments/assets/e90a1874-4226-4490-bede-9723998000bd"
/>

3* The url is taken from the `url` attribute of the provenance. Note
there are other choices, such as `sourceDataUrl` which is sometimes the
same (as in OECDRegionalDemography_LifeExpectancy) but sometimes
different (as in EurostatData_Fertility, where the `sourceDataUrl` is a
direct download link.)

5* The secondary name of the source is not found in the screenshot, but
is taken from "isPartOf[0]->name".

10* This comes directly as a DCID and so is looked up, with the
description being used.

12* This is looked up. We take the absoluteName of the last item in the
stat var's path (which will be "Demographics", "Economics", etc.

### Citation

The citation is a single, united citation for all sources. The text
"with minor processing by Data Commons" is appended to the end of the
citation.

The action of copying is indicated to the user by swapping out the icon.
We can consider adding in "toast" (such as the ones used in Gmail) to
give further feedback to the user.

## Codacy

Codacy produced the following note:

```
expected parameter: 'omission' to have a typedef
```

However adding in the type definition created an opposite linting error
because the type definition can be inferred.

---------

Co-authored-by: Pablo Noel <pablo@next-solutions.ca>