Metadata Modal Error State (#5189)
## Description

To account for the possibility of unforeseen errors in the metadata
fetching, we are now providing error feedback to the user. We catch
errors thrown from `fetchMetadataModal` and use those to display an
error message in the modal.

<img width="780" alt="metadata_modal_error"
src="https://github.com/user-attachments/assets/832927f7-8f9c-4e3a-b0d5-fb70ffd4f866"
/>

## Testing

The easiest way to test the metadata modal error state is to
artificially trigger an error by adding a line to throw one in the
fetchMetadata function itself.

- Open the file `static/js/tools/shared/metadata/metadata_fetcher.ts`.
- At the top of the function `fetchMetadata` add the following line:
- `throw new Error("test");`
- Once the front-end has built, refresh and load any metadata dialog on
any chart and you will see the error state.