This refactoring standardizes the data-sourcing pattern for the FacetSelector component.
Previously, parent chart components would create a promise and pass it into the FacetSelector for resolution.

Now parent components are now responsible for resolving their own data-fetching promises, facilitated by a new generic usePromiseResolver hook.

The goal of this change is to make the data sourcing of the facet selector more standard, in order to make each component easier to maintain and test.

In particular, this was done in mind with the upcoming task to introduce more complex metadata into the facet selector.