Dataset Selector Modal V0 (#5204)
## Description

This PR is a follow up to `Dataset Selector V0`, which implemented the
conversion of the chart header and button.

This follow up implements version `V0` of the dataset selector dialog,
as found in the [slide
deck](https://docs.google.com/presentation/d/1ztyZwu8b3YzU1MMAFPI3iC-Hm00CE63rPb7eP6vGZlw/edit?pli=1&slide=id.g347cc273f35_0_120#slide=id.g347cc273f35_0_120),
page 16 as of the creation of this PR.

As per discussion, along with the layout updates from the deck, the
accordions in the old facet selector have been removed, and we are using
the DC dialog and button components throughout.

For charts with single stat vars, the copy of the button and dialog
title is "Select a dataset". For multiple stat vars, the copy of the
button and the dialog title is "Select datasets". This reflects the
multi-stat-var example in the slide deck on page 17

The design of the multi-stat-var version is also from that 

## Screenshots

#### Single var:

<img width="598" alt="Screenshot 2025-06-11 at 6 26 26 AM"
src="https://github.com/user-attachments/assets/869d32e3-38be-4ad2-a870-88c1a372abbc"
/>

#### Multi var:

<img width="621" alt="Screenshot 2025-06-11 at 6 25 56 AM"
src="https://github.com/user-attachments/assets/85eb08e4-0358-433b-b365-4738c77e930b"
/>

## Notes

To keep consistency with other areas, for the "Update" button, we used
the action variant with the border, rather than the borderless version
(the dialogs in the slide deck varied in their approach).

One this PR is merged, we will follow up with V1/V2.

## Testing

Single stat var:


http://localhost:8080/tools/visualization#visType%3Dtimeline%26place%3Dcountry%2FSWE%26placeType%3DAdministrativeArea1%26sv%3D%7B%22dcid%22%3A%22LifeExpectancy_Person%22%7D

Multiple stat vars:


http://localhost:8080/tools/visualization#visType%3Dtimeline%26place%3DgeoId%2F06___geoId%2F08%26placeType%3DCounty%26sv%3D%7B%22dcid%22%3A%22Median_Age_Person%22%7D___%7B%22dcid%22%3A%22Count_Person_Female%22%2C%22pc%22%3A%221%22%7D___%7B%22dcid%22%3A%22Count_Person_Male%22%2C%22pc%22%3A%221%22%7D

---------

Co-authored-by: Pablo Noel <pablo@next-solutions.ca>