Updated the facet selector so that it will not show until it has options to display. This was required because the way that the new and old tools resolve these promises is very different.
This would be solved a refactor of the data fetching to a more standard pattern where the promises are resolved outside of the button and only the final result is passed into the button (effectively what is happening in the old tool, in a roundabout way).

A todo has been added for that.