Update to the ga_events tests to do the following:
Each entry point is now wrapped in the theme provider.

The tests that used to open the facet selector dialog were complicated by the fact that now the dialog can only be opened if we have facets to show. These were difficult to mock within the individual tests (because of intervening with the mocks affected the chart), and so we removed that test from each of the charts, and instead added a new test that tests the facet selector dialog in isolation (allowing us to properly mock it).

We also updated the snapshots.