Add option to generate metadata in French and Spanish (#5211)
Adds the ability to generate metadata for stat vars in French and
Spanish by passing in a new feature flag "-language" in the script. As
part of this change I refactored some functions in the script to return
the metadata as a list of dictionaries rather than a list of
StatVarMetadata objects. This was done to standardize the format used
regardless of whether the user was generating sentences or not, and in
different languages, as sticking with StatVarMetadata objects could only
support metadata generation in English, and would result in inconsistent
output formats depending on whether sentences were being generated by
Gemini or not.