[Chart Injection] UI Tweaks - Border fix, place overview tile removal and per capita checkbox control (#5206)
See the separator now between the highlight and main
SubjectPageMetadatas: https://screenshot.googleplex.com/4KgkVYTkAghN33s
-- Verfiied that we didnt add another separator at the very bottom.

Removes the place overview tile from ever being rendered in the
renderTiles function. Note that in the past it was shown but we have
since implemented redirection to the place page whenever it would have
shown up.

Updated the block controls to only render per capita checkbox if there
is a denominator.
Remove denominator and snapToLatest option for highlight results.