[Chart Injection] Remove the map view from pages with an injected chart. (#5234)
This can happen today for multiple reasons:
- Some stat vars are incorrectly not associated to topics - will be
resolved
- Even when that is resolved, it's possible for there to only be data on
the injected chart, and not on any other stat var, therefore we need to
account for this use case.


If the main page metadata only contains an Overview component (i.e. the
map), we use the highlight page metadata as the main page metadata.
- Before: https://screenshot.googleplex.com/6jT5yLQQQWKXDNa
- After: https://screenshot.googleplex.com/7k6NN2EpSt65TRt

Note the topic is still missing in the header but that will be resolved
once we have the topic association!