This push implements the beginning of a set of updates to address the fact that: - sometimes a data selector has one or more stat vars that do not have a selectable option (i.e., the stat var contains only one facet). - sometimes a data selector contains no stat vars at all that have a selectable option.
In the case where a data selector contains no stat vars with selectable options, the button is replaced with text to indicate this.

In the case where only some of the stat vars have no selectable options, the data selector can be opened, but those particular stat vars only contain a view version with no radio button.