[Explore] Add Overall Explanation Endpoint (#5277)
Added the Flask route to serve the generated explanation and the helper
functions that call Gemini to generate the explanation. Most likely, the
Gemini configurations such as the prompt are not final but they can be
easily tweaked and it allows us to continue with the rest of the
feature's logic.