Block-level dataset selector (line, highlight, scatter) (#5269)
## Description

This PR implements dataset selection at the level of the blocks used in
the NL search.

This first round applies it to the line, highlight and scatter charts.
These are the three charts for which there are no changes required in
the dataset selector mechanism itself

These updates are limited to the NL search, and exclude the web
components.

### Updates to the tiles

We introduced a per-block cache that memoizes StatVarSpec[] results.
This prevents block-level re-renders from
triggering refetches in the child.

This update was prompted by the additional rendering cycles from the
facet selector handling. However any similar functionality added into
the block would have required this as well.

Although we tried to keep the changes in the block itself (to minimize
the breadth of the PR), we did have to update some aspects of the tiles
themselves. For example, the highlight and ranking tiles were triggering
a fetch on a dependency of [props] - resulting in duplicate calls. We
now destructure the props and only trigger the fetch if a relevant prop
has changed.

This overall sensitivity that the tiles have to re-rendering could be
hardened in a future PR (thus allowing the tiles to be used in any
context without rendering causing unnecessary fetches).

## Screenshots

### Selector on a line/highlight pair

<img width="1369" alt="Screenshot 2025-07-07 at 9 44 55 AM"
src="https://github.com/user-attachments/assets/4727a1c7-5c3c-472e-b63c-31e566ef7e9c"
/>

### Selector on a scatter plot

<img width="1364" alt="Screenshot 2025-07-07 at 9 44 04 AM"
src="https://github.com/user-attachments/assets/8c35c0a5-9b51-4190-9397-e513c614e11a"
/>

### Block with only one dataset available

<img width="1366" alt="Screenshot 2025-07-07 at 9 44 34 AM"
src="https://github.com/user-attachments/assets/0df13acd-8120-4851-a1dd-67c2adaa7a04"
/>


## Testing

Between the following calls, you will see examples of each of the three
chart types involved in this PR:

### Line, highlight


http://localhost:8080/explore#q=life+expectancy+in+sweden&client=ui_query

### Scatter

http://localhost:8080/explore#q=poverty+vs+employment&client=ui_query

### No dataset choices available

http://localhost:8080/explore#q=united+states+statistics&client=ui_query

## Feature Flag

This functionality is behind the same feature flag used for the rich
metadata modal and the rich facet selector, and so will show up locally
and in autopush.

---------

Co-authored-by: Pablo Noel <pablo@next-solutions.ca>