Mobile Header with no search bar (#5310)
## Description

This PR addresses an issue with the height of the header in mobile mode
when search is off.

When the search bar is displayed, on smaller mobile screens the search
bar displays below the rest of the header (resulting in a taller
header).

When the search bar is not displayed, there was extra space under the
header that this PR resolves.

## Screenshots

Before:

<img width="297" height="89" alt="Screenshot 2025-07-18 at 2 49 36 PM"
src="https://github.com/user-attachments/assets/23a1e091-d140-4ab9-b942-062e94d3bbd5"
/>

After:

<img width="287" height="117" alt="Screenshot 2025-07-18 at 2 50 53 PM"
src="https://github.com/user-attachments/assets/d269c691-e701-4984-9a15-dbb1c85d7796"
/>

## Testing

To test this, turn on the `standardized_vis_tool` flag in `local.json`
and restart the flask server. This will cause the old vis tools to turn
off the header view.

The mobile header should now be a consistent and appropriate height
whenever we are in a mobile size (i.e., the mobile menu rather than the
desktop menu show).

When viewing a page that has a search bar, the header will behave
exactly the same as in master.

Co-authored-by: Pablo Noel <pablo@next-solutions.ca>