This push significantly overhauls facet handling for bar charts (and any other charts that later are flagged as allowed grouped facet selections). The facet selector is now separated into multiple files, and multiple components, with the primary division being grouped versus standard stat var handling.
When stat var handling is standard, the dialog will display as it always did. When they are grouped, the dialog will no longer render separate facet selection sections for the different stat vars, but rather render a single list for all the stat vars, containing every facet available for any of the stat vars.

If any of those facets are not available for all stat vars, this is indicated.