[Vis Tools] Code Cleanup: refactor switch tool button (#5338)
Update the "switch tool version" button to use the shared button
component instead of custom anchor styling. This allows for the button
to follow set theming and standardizes the look with the rest of the
buttons on the site.

Screenshot:
<img width="1461" height="650" alt="image"
src="https://github.com/user-attachments/assets/33100f6c-25e8-4463-a0ea-aed466dd420b"
/>