Updates to the Block component to implement facet selection capability for maps.
In addition to turning on facet selection for the maps, this contains updates to the snap-to-highest checkbox functionality both to allow for facet's to be incorporated into the decision, and to trigger a reload of the check when a facet changes.

In making these changes, I also did some refactoring of the highlight facet aspect of eligibility. We are now checking for that directly in the function that checks for other aspects of snap-to-highlight eligibility.

The need to incorporate facets into the decision-making for snap-to-highlight required that highestCoverageDatesEqualLatestDates be updated to use the Data Commons Web Client rather than the Data Commons Client, as the latter does not have facet selection capacity.