Update to the test_map_select_different_facet test to future-proof it from failure due to data ingestion.
Future data ingestion could either:
- change the dates on either side of the toggle
- make it so that the toggle is no longer available.

To prevent the change of dates from breaking the test, we no longer hard-code the expected dates, and instead check that the date changed from a single date (checked) to a date range (unchecked).

To alleviate the second situation, we check if the checkbox can be toggled, and if it can't, it means future data ingestion has made it so that the highest coverage data is the latest data. In this case, the app is not wrong, and the test should not fail (but we emit a warning to alert us that we need to review the test).