There appears to sometimes be a discrepancy between the sources list causing an issue with the bar chart facet selector test.
Locally, on the live site and in staging, the sources list is "Sources: census.gov, data.census.gov • Show metadata".

However the test (not local) is finding "Sources: census.gov, data.census.gov • Show metadata".

The cause of the discrepancy should be reviewed, but this push updates the bar facet selector test to test for the expected sources individually in a way that will pass both versions.